Create Rotated Search
An Array with all the elements sorted across a pivot is given to you, You need to search the given element in the sorted rotated array. If it's not found print -1 , otherwise print the index of that element.

Input Format
An integer n - Size of the Array followed by elements of that array and an integer X to be searched.

Constraints
1 <= n <= 10^6 1 <= arr[i] <= 10^9 , where arr[i] is any element in the array.

Output Format
An integer index of that element or -1 , if not found.

Sample Input
5
5 6 1 2 3
1
Sample Output
2
Explanation
Index of element 1 is 2.